Rename command and fix tenantcluster
