+++ cmd/action/verify/cluster/ha/error.go
@@ -1,4 +1,4 @@
-package hasetup
+package ha
 
 import "github.com/giantswarm/microerror"
